- 删除mplayerx插件 - 删除一些文件 - SoraParticleSystem添加boundingBox支持 - SoraCore添加renderBox, 画一个框 - 修正所有颜色格式为uint32 - 修正RTTI一些说明 - SoraPlatform内去掉DWORD的定义 - SoraFileUtility添加getFileName - SoraFileUtility的readFiles函数名修正为readFileWithSize - SoraLocalizer添加对resource的localize支持, 格式为string = string, 即"xxx.png" = "xxx_chn.png", 方便内置文字资源的本地化, - SoraOGLRenderer添加flushTrip, flush模式为GL_TRIANGLES

+++ Sora/SoraPlatform.h
@@ -112,7 +112,7 @@ typedef		double			double32;
 typedef		long			long32;
 typedef		unsigned long	ulong32;
 
-typedef unsigned long       DWORD, ULONG;
+typedef unsigned long       ULONG, ulong;
 typedef unsigned short      WORD;
 typedef unsigned char       BYTE;
 typedef unsigned int		UINT;
